Added support for predefined traces.
A first predefined trace was added, which traces
changes to the system properties. The logging is
done via converters. For this to work, the converters
had to be included in the bootstrap class path. This
was done by creating a seoond jar containing only the
files need to be in the bootstrap class path. This jar
will then be added to the bootstrap classpath by the
agent at runtime.